Create a bucket list.

A bucket list can contain more than one rank_list and allows drag-and-drop of items between the different lists.

bucket_list(header = NULL, ..., group_name, group_put_max = rep(Inf,
  length(labels)), selector = NULL, options = sortable_options(),
  style = css_bucket_list())

Arguments

header
Text that appears at the top of the bucket list. (This is encoded as an HTML <p> tag, so not strictly speaking a header.)

...
One or more specifications for a rank list, and must be defined by add_rank_list.

group_name
Passed to sortable.js as the group name

group_put_max
Not yet implemented

selector
This is the css id to use, and must be unique in your shiny app. If NULL, the function generates a selector of the form rank_list_id_1, and will automatically increment for every rank_list.

options
Options to be supplied to sortable_js object. See sortable_options for more details

style
A css stylesheet, provided as a character string. See also css_rank_list().

Examples

## Example of a shiny app
if (interactive()) {
  app <- system.file("shiny-examples/bucket_list/bucket_list_app.R", package = "sortable")
  shiny::runApp(app)
}
## -- example-bucket-list ---------------------------------------------

## bucket list

bucket_list(
  header = "This is a bucket list. You can drag items between the lists.",
  add_rank_list(
    text = "Drag from here",
    labels = c("a", "bb", "ccc"),
    input_id = NULL
  ),
  add_rank_list(
    text = "to here",
    labels = NULL,
    input_id = NULL
  )
)

## bucket list with three columns

bucket_list(
  header = c("Sort these items into Letters and Numbers"),
  add_rank_list(
    text = "Drag from here",
    labels = sample(c(1:3, letters[1:2])),
    input_id = NULL
  ),
  add_rank_list(
    text = "Letters",
    input_id = NULL
  ),
  add_rank_list(
    text = "Numbers",
    input_id = NULL
  )
)